+++ library/vector.cpp
@@ -16,16 +16,16 @@ vector_t vec_negate(vector_t v) {
     return output;
 }
 
-vector_t vec_multiply(double scalar, vector_t v) {
+vector_t vec_multiply(float scalar, vector_t v) {
     vector_t output = {scalar * v.x, scalar * v.y};
     return output;
 }
 
-double vec_mag(vector_t v1) {
+float vec_mag(vector_t v1) {
 	return sqrt(v1.x * v1.x + v1.y * v1.y);
 }
 
-double vec_distance(vector_t v1, vector_t v2) {
+float vec_distance(vector_t v1, vector_t v2) {
 	return vec_mag(vec_subtract(v1, v2));
 }
